Fixes to rpUserCtrl.
Do not load tabs until we've loaded the user and set isMe. This way we
avoid setting the tab incorrectly if we try to load restricted tab
before we've allowed it.
Do not set the title after the posts have loaded to the author of the
first post. Set the title from the routeParams when initiate controller
and it should be enough.