disabled getInstance() in redditAuthHandler, no longer necessary.
But kept refreshToken. Server manages user refreshTokens based on
generatedIds. A different refreshToken is used for a different device
and stored in the database under the same userID. The server needs a
Snoocore object to get the userId. Otherwise we could just disable the
whole thing. Also needs to deauth and remove the token from the database
when the user logs out.

There is probably a simpler way of doing this now that we don't need to
maintain snoocore objects and retrieve them all the time etc.