Fix user authentication problems.
Problem was the refresh token not being found in the database.

Change all database queries to look up the user first and then use the
generatedState to find the refreshToken.

If the user or the refreshToken is not found during getInstance the user
is redirected to logout.

Also change the query in rpSettingsHandler to only try and find the user
by userId.

I'm not sure if there will be problems getting the user by only id, we
used to do it that was but then changed to start using the userId and
generatedState.